Struct smallstr::SmallString

pub struct SmallString<A: Array<Item = u8>> { /* fields omitted */ }

A String-like container that can store a small number of bytes inline.

SmallString uses a SmallVec<[u8; N]> as its internal storage.

Implementations

impl<A: Array<Item = u8>> SmallString<A>

pub fn new() -> SmallString<A>

Construct an empty string.

pub fn with_capacity(n: usize) -> SmallString<A>

Construct an empty string with enough capacity pre-allocated to store at least n bytes.

Will create a heap allocation only if n is larger than the inline capacity.

pub fn from_str(s: &str) -> SmallString<A>

Construct a SmallString by copying data from a &str.

pub fn from_string(s: String) -> SmallString<A>

Construct a SmallString by using an existing allocation.

pub fn from_buf(buf: A) -> Result<SmallString<A>, FromUtf8Error<A>>

Constructs a new SmallString on the stack using UTF-8 bytes.

If the provided byte array is not valid UTF-8, an error is returned.

pub unsafe fn from_buf_unchecked(buf: A) -> SmallString<A>

Constructs a new SmallString on the stack using the provided byte array without checking that the array contains valid UTF-8.

Safety

This function is unsafe because it does not check that the bytes passed to it are valid UTF-8. If this constraint is violated, it may cause memory unsafety issues, as the Rust standard library functions assume that &strs are valid UTF-8.

pub fn inline_size(&self) -> usize

The maximum number of bytes this string can hold inline.

pub fn len(&self) -> usize

Returns the length of this string, in bytes.

pub fn is_empty(&self) -> bool

Returns true if this string is empty.

pub fn capacity(&self) -> usize

Returns the number of bytes this string can hold without reallocating.

pub fn spilled(&self) -> bool

Returns true if the data has spilled into a separate heap-allocated buffer.

pub fn drain(&mut self) -> Drain<'_>

Empties the string and returns an iterator over its former contents.

pub fn push(&mut self, ch: char)

Appends the given char to the end of this string.

Examples

use smallstr::SmallString;

let mut s: SmallString<[u8; 8]> = SmallString::from("foo");

s.push('x');

assert_eq!(s, "foox");

pub fn push_str(&mut self, s: &str)

Appends the given string slice to the end of this string.

Examples

use smallstr::SmallString;

let mut s: SmallString<[u8; 8]> = SmallString::from("foo");

s.push_str("bar");

assert_eq!(s, "foobar");

pub fn pop(&mut self) -> Option<char>

Removes the last character from this string and returns it.

Returns None if the string is empty.

pub fn grow(&mut self, new_cap: usize)

Reallocates to set the new capacity to new_cap.

Panics

If new_cap is less than the current length.

pub fn reserve(&mut self, additional: usize)

Ensures that this string’s capacity is at least additional bytes larger than its length.

The capacity may be increased by more than additional bytes in order to prevent frequent reallocations.

pub fn reserve_exact(&mut self, additional: usize)

Ensures that this string’s capacity is additional bytes larger than its length.

pub fn shrink_to_fit(&mut self)

Shrink the capacity of the string as much as possible.

When possible, this will move the data from an external heap buffer to the string’s inline storage.

pub fn truncate(&mut self, len: usize)

Shorten the string, keeping the first len bytes.

This does not reallocate. If you want to shrink the string’s capacity, use shrink_to_fit after truncating.

Panics

If len does not lie on a char boundary.

pub fn as_str(&self) -> &str

Extracts a string slice containing the entire string.

pub fn as_mut_str(&mut self) -> &mut str

Extracts a string slice containing the entire string.

pub fn clear(&mut self)

Removes all contents of the string.

pub fn remove(&mut self, idx: usize) -> char

Removes a char from this string at a byte position and returns it.

Panics

If idx does not lie on a char boundary.

pub fn insert(&mut self, idx: usize, ch: char)

Inserts a char into this string at the given byte position.

Panics

If idx does not lie on char boundaries.

pub fn insert_str(&mut self, idx: usize, s: &str)

Inserts a &str into this string at the given byte position.

Panics

If idx does not lie on char boundaries.

pub unsafe fn as_mut_vec(&mut self) -> &mut SmallVec<A>

Returns a mutable reference to the contents of the SmallString.

Safety

This function is unsafe because it does not check that the bytes passed to it are valid UTF-8. If this constraint is violated, it may cause memory unsafety issues, as the Rust standard library functions assume that &strs are valid UTF-8.

pub fn into_string(self) -> String

Converts the SmallString into a String, without reallocating if the SmallString has already spilled onto the heap.

pub fn retain<F: FnMut(char) -> bool>(&mut self, f: F)

Retains only the characters specified by the predicate.

In other words, removes all characters c such that f(c) returns false. This method operates in place and preserves the order of retained characters.

Examples

use smallstr::SmallString;

let mut s: SmallString<[u8; 16]> = SmallString::from("f_o_ob_ar");

s.retain(|c| c != '_');

assert_eq!(s, "foobar");